MGDOBR-384: use BaseConnectorAction
