+++ manager/src/main/resources/application.properties
@@ -93,6 +93,8 @@ managed-connectors.auth.client-id=${MC_AUTH_CLIENT_ID:fakeClientId}
 managed-connectors.auth.token-path=${MC_AUTH_TOKEN_PATH:fakeTokenPath}
 managed-connectors.auth.offline-token=${MC_AUTH_OFFLINE_TOKEN:fakeOfflineToken}
 
+managed-connectors.log-enabled=true
+
 # RHOAS
 rhoas.timeout-seconds=60
 rhoas.ops-account.client-id=fakeClientId
